// client/messages.ts
/** this file is automatically generated, do not edit **/

export type ServerMessage = { type: "Update"; store: string; value: string };
export type ClientMessage = { type: "Set"; store: string; value: string } | { type: "Get"; store: string } | { type: "Subscribe"; store: string; initial: string } | { type: "Unsubscribe"; store: string };

export type ClientMessageTypes = ClientMessage["type"];
export type ServerMessageTypes = ServerMessage["type"];
export type ServerMessageMap<T extends ServerMessageTypes> = Omit<Extract<ServerMessage, { type: T }>, 'type'>;
export type ClientMessageMap<T extends ClientMessageTypes> = Omit<Extract<ClientMessage, { type: T }>, 'type'>;